backend/requirements.txt
Flask
Flask-CORS
Werkzeug
vellum-ai
python-dotenv 
requests
google-genai
numpy
supabase